moving methods to a separate package
